chore: little fixes
