+++ src/types/database/index.ts
@@ -111,3 +111,16 @@ export interface Auth {
   token: string
   lastDate: string
 }
+
+export interface ResponseLogin {
+  id: number
+  spoId?: number
+  groupId: number
+  login: string
+  phone: string
+  birthday: string
+  firstName: string
+  lastName: string
+  middleName: string
+  token?: string
+}
