+++ src/types/index.ts
@@ -1,3 +1,2 @@
 export * from './database'
-export * from './database/returnTypes'
 export * from './converterTypes'
